3ds-vla/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
torch==2.0.0+cu117
torchvision==0.15.1+cu117
fairscale
sentencepiece
Pillow
opencv-python
gradio
tqdm
# git+https://github.com/csuhan/timm_0_3_2.git
# git+https://github.com/openai/CLIP.git